01/04

+++ Project/Engine/CMaterial.cpp
@@ -148,7 +148,7 @@ void CMaterial::UpdateData()
 	const_buffer->SetPipelineStage(PIPELINE_STAGE::PS_ALL);
 	const_buffer->UpdateData();
 
-	for(size_t i =0; i<texture_array_.size(); ++i)
+	for(UINT i =0; i<(UINT)texture_array_.size(); ++i)
 	{
 
 		if (nullptr != texture_array_[i])
